requirements.txt
pylint==2.3.1
requests==2.21.0
django==2.2
djangorestframework==3.9.3